feat: integrate the auth to each endpoint for security

+++ app/app.py
@@ -1,6 +1,6 @@
 from fastapi import FastAPI, HTTPException, File, UploadFile, Form, Depends
 from app.schemas import PostCreate, UserCreate, UserRead, UserUpdate
-from app.db import Post, create_db_and_tables, get_async_session
+from app.db import Post, create_db_and_tables, get_async_session, User
 from sqlalchemy.ext.asyncio import  AsyncSession
 from contextlib import asynccontextmanager
 from sqlalchemy import select
@@ -65,6 +65,7 @@ app.include_router(fastapi_users.get_users_router(UserRead, UserUpdate), prefix=
 async def upload_file(
     file: UploadFile = File(...),
     caption: str = Form(""),
+    user: User=Depends(current_active_user),
     session: AsyncSession = Depends(get_async_session) # dependency injection to get an async DB session before func can run
 ):
     temp_file_path = None
@@ -85,6 +86,7 @@ async def upload_file(
 
         post = Post(
             caption=caption,
+            user_id=user.id,
             url=upload_result.url,
             file_type="video" if file.content_type.startswith("video/") else "image",
             file_name=upload_result.name
@@ -105,7 +107,8 @@ async def upload_file(
 
 @app.get("/feed")
 async def get_feed(
-    session: AsyncSession = Depends(get_async_session)
+    session: AsyncSession = Depends(get_async_session),
+    user: User=Depends(current_active_user),
 ):
     result = await session.execute(select(Post).order_by(Post.created_at.desc())) #can do .filter_by for more filtering etc.,
     # result var is of type cursor object
@@ -116,17 +119,23 @@ async def get_feed(
         posts_data.append(
             {
                 "id": str(post.id),
+                "user_id": str(post.user_id),
                 "caption": post.caption,
                 "url": post.url,
                 "file_type": post.file_type,
                 "file_name": post.file_name,
-                "created_at": post.created_at.isoformat()
+                "created_at": post.created_at.isoformat(),
+                "is_owner": post.user_id == user.id,
+                "email": post.user.email
             }
         )
     return {"posts": posts_data}
 
 @app.delete("/posts/{post_id}")
-async def delete_post(post_id: str, session: AsyncSession = Depends(get_async_session)):
+async def delete_post(
+    post_id: str,
+    user: User=Depends(current_active_user),
+    session: AsyncSession = Depends(get_async_session)):
     try:
         post_uuid = uuid.UUID(post_id)
 
@@ -135,6 +144,9 @@ async def delete_post(post_id: str, session: AsyncSession = Depends(get_async_se
 
         if not post:
             raise HTTPException(status_code=404, detail="Post Not Found")
+        
+        if post.user_id != user.id:
+            raise HTTPException(status_code=403, detail="You do not have permission.")
         
         await session.delete(post)
         await session.commit()
